multiply-me/requirements.txt
google-genai
python-dotenv
